Run data binding metrics task as nonblocking
This task can be long-running due to the number of files it has to
process, and potentially due to long-running file I/O operations.
Addressing that by making the read action non-blocking, and reducing the
scope to only what actually needs to be in the read action.

The call to FileTypeIndex.getFiles has process cancelation checks in it,
and I've added an explicit check for each call made to the
BindingXmlIndex.

Fixes: 312754297
Test: DataBindingTrackerTest.testDataBindingPollingMetadataTracking (and
others)

Change-Id: Ib2a7f4c375890bc078c18b11fae6137718130945