[lint] Make Groovy visitors consistent over plugins Dsl
When inspecting a method call or application statement within a
closure, we might end up with something we don't fully
understand (e.g. a set of chained method calls).  In that
circumstance, we don't want to deal with the right-most call, but
rather the left-most call that we do understand.  For example, in

    plugins {
      id 'android' version '2.3.3' apply true
    }

we want to check the call `id 'android'` on the `plugins` block,
rather than the `apply true` call on whatever the return value of `id
'android' version '2.3.3'` is.

That means we need to be able to determine whether we can actually
resolve all the qualifiers of a given call, and not just return up to
three of them.  Define a nullable variant of
getReferenceExpressionNames, and use a null return to detect whether
we should recurse into subexpressions in
handleApplicationOrMethodCallInClosure().

Bug: 279885639
Test: added
Change-Id: I6186341202bb6e404057bef0b61e0a0c60c38b93