[PSD] Guava version number handling in Dependencies tab
Some guava artifacts encode aspects of their target platform in their
version number.

Really.

So if a user depends on e.g. "com.google.guava:guava:32.1.3-android",
they might be surprised to find that the PSD suggests that they
upgrade this dependency to "com.google.guava:guava:32.1.3-jre".

When looking for values for a dependency version property, if the
group is `com.google.guava`, and the current version ends with `-jre`
or `-android`, restrict the values offered to only those with versions
matching the suffix.

Bug: 315310265
Test: manual
Change-Id: I9dfb17ffce044bcf78c7624b3c06b43c04d5994b