Fix string ordering for translations editor
Strings in the translations editor were intended to display in the same
order as they were entered into strings.xml by default (at least within
a single file). This was originally implemented with a few changes,
notably:

 - I8f0c80e6d9cc3aac3bfe28f2b658f307b0254732
 - If53d2de2c14f753fc28b6c482d57f67ddd89f08f

The test AppResourceRepositoryTest.testStringOrder verified the
behavior of the repository, in that it should returns strings in order.
However the use of fake repositories caused it to exercise a different
code path in the super MultiResourceRepository class than what happens
in real code. (This is because the AppResourceRepository in the test had
only one child, but in real code with a strings file it always has at
least two children due to sample data.)

The change to use the custom PerConfigResourceMap in
MultiResourceRepository broke the ordering, unfortunately. But
fortunately, it's easy to add ordering back in by using an ordered map
implementation internally.

Bug: n/a
Test: AppResourceRepositoryTest.testStringOrder
Change-Id: Ia35d79c7c5b225ea5dbd5051fb7092290114dc04