Dispose repository children outside of locks
When `StudioResourceRepositoryManager` resets its resources, it disposes
the old children that implement `Disposable`. This change moves the
dispose calls outside of resource locks; the locks only need to guard
access to the current set of repository children that are stored. The
old objects can be safely disposed outside of these locks, which will
also help prevent deadlocks if any of the dispose calls end up being
non-trivial.

Bug: 313309846, 304316363
Test: StudioResourceRepositoryManagerTest.testDisposal
Change-Id: I03950dcc1d127eeca7f01abeeb3edf3afc8e006b