Improve display name for initial AVD
The AVD that we create in the first run wizard has a different name
format than what we default to in the normal AVD creation flow, e.g.
"Pixel_3a_API_34_extension_level_7_x86_64". We shouldn't be using
underscores in the display name, we don't need the ABI, and spelling
out "extension level" is unnecessary.

This changes the display name to something like "Pixel 3a API 34-ext7"
and unifies the code with ConfigureAvdOptionsStep's updateComponents().

Bug: 315946845
Test: AndroidVirtualDeviceTest
Change-Id: Idae7426527dbdfcb74676260e44dcecd89ebcef4