Remove NDK selection from project structure
This dropdown is from the pre-side-by-side-NDK days where we only had
ndk-bundle, and this dialog was how the user told us where their "custom"
NDK was on the disk.

Since AGP 4.1 and newer, we have have 'ndkPath' Gradle DSL property that
controls this. Hence, this dialog is obsolete.

For pre AGP 4.1 users, I wanted to keep this UI and fix minor issues
around usability. But, I noticed that the dialog has been completely
broken in multiple different ways:

1. Auto-download downloads a non-stable ndk-bundle at r21-beta04.
2. After auto-download, the dropdown points to side-by-side NDK at the
   AGP default (r25 at the moment), not the downloaded NDK.
3. After auto-download, the dropdown does not put the downloaded NDK
   to the list of NDKs for the user to select.
4. Irrespective of which NDK you select, when you hit OK/Apply, the
   NDK validation logic kicks-in and rejects the NDK you selected,
   because it has a bug that expects the NDK to have a "platforms"
   subdirectory, yet NDK doesn't have such a directory.

And this has been the behavior for 3+ years. It's better to drop
these rather than trying to fix them.

Bug: 314821739
Test: manual, verify link does not trigger
Change-Id: I76a0db3f82f2b6160489c7964c68e563b0d8a749