Fix resource repository disposable registrations
Currently a number of resource repositories that are disposable are
registered as such following their construction, but not as part of the
construction. However, during their construction they are registered as
the parent disposable of a memory watcher. In practice, this means the
following happens:

1. Repository 'a' is created.
2. Repository 'a' is registered as the parent disposable of a memory
   watcher. Since 'a' is not in the disposal graph, it is registered as
   a root.
3. Repository 'a' is registered as a child disposable of some parent
   'p'. It is removed from being a root.

Flakiness in TransitiveTestRClassesTest#testResolveScope exposed a race
condition. After step 2 above, the process is canceled. Repository 'a'
is left as a dangling root element.

This registration pattern has been around for a while; it was unlikely
this race condition was happening because not much logic was involved in
constructing the repositories. (Still possible, though; there are
multiple places in construction that ProcessCancellationException can be
thrown.) The recent change ag/25629715 made this much more common,
though, leading to test flakiness.

I'm fixing this by putting in place the much more idiomatic pattern with
disposables, whereby the parent is passed into the constructor and
registered there. That means it's registered *before* the child memory
watcher is registered, and then the repository never goes through a
stage of being a root disposable.

Test run passing 500x:
https://fusion2.corp.google.com/invocations/cd36e78a-c037-4ccd-bb00-d93850889f25

Test: TransitiveTestRClassesTest#testResolveScope
Fixes: 316186264
Change-Id: Ifbfc5d67309c2a46aa6f35c481e4120099bc0d63