Fix HostPanel initialization
The previous code relied on a waitForSmartMode call before starting the
reference search. This works most of the times but there are no
guarantees that dumb mode will not start again before the search begins.
This change ports the code to use the new coroutines flow which handles
retries of reads and smart mode.

Test: HostPanelTest
Fixes: 316185667
Change-Id: I5393429a243fd5128e35ccede4a99c4293acb47f