Update ModuleResourceRepository disposal
DynamicValueResourceRepository and ModuleResourceRepository have some
circular references that make disposal registration difficult. This
change attempts to ensure that both are registered with a parent
disposable during construction, to avoid leaks.

(The existing code handled this with a lot of try/catch/rethrows, and it
looks like it was solid; this is more of a defensive change to codify
behavior and prevent any bugs from creeping in as things change in the
future.)

To accomplish this, I've moved the logic figuring out child repositories
into the initialization of the repo, rather than being calculated
outside and passed in. In doing that, it became clear that
forMainResources and forTestResources are identical except for how they
handle dynamic resources (which is easily distinguished by a bool
parameter), and so those methods have now been collapsed as well.

Bug: n/a
Test: n/a
Change-Id: I5fb2690d245981d93b1ca57f1e9cd0f687132949