Fix the flaky test: testResetRecompositionCounts
Recently a columnUpdate check was added in an attempt to fix this
flaky test.

Unfortunately there is a delayed columnUpdate event sent after the
tree is initially populated. It is possible (but hard to reproduce)
that the columnChange detector was setup before the previous column
update happened, thus rendering the wait after the click uneffective.

Fix this.

Fixes: 311270338
Test: LayoutInspectorTreePanelTest.testResetRecompositionCounts
Change-Id: Ibc1194f40e22892bf385987cba2a1af284bfeb7d