[SDK Index] Show all issues individually in PSD
...instead of grouping them when two or more are present.

When two or more SDK index issues are present for the same library
version we show a generic message that says that two or more issues are
present, without displaying what issues there are. After this change we
no longer group all issues into a single one and instead display one
message for each issue.

Bug: 316038712
Test: PsAnalyzerDaemonKtTest, PolicyIssuesPsdTest, SdkIndexPsdTest
Change-Id: Idd9201e66268288e95d361ae908f3f0fd18a9526