Add App Language selector to UI settings panel
This setting controls the Language for an application, which is
different from the rest of the UI settings which are device settings.

The application is selected from the android applications in the
current project (using the current run configuration as selector if
multiple apps exist in the project).

Only the main Android apps in the current project can be changed.
The UI will provide a dropdown of all apps that have multiple
locales, and supply an app locale selector for the currently
selected app.

The available locales are found from the project sources since there
is no ABD command for getting it from the device.
Such a command is requested here: b/317228996.

Bug: 308982261
Test: Added unit tests
Change-Id: I5702f27e6a5ff926b3579593b1027045b3938bd9