Disable an assert in AddComposableAnnotationQuickFixTest
...since Compose compiler 1.5.4 reports COMPOSABLE_EXPECTED error
on NonComposableFunction while it should only do this for a lambda
taking @Composable as an argument.

This scenario is not common because the only way to use a composable
inside of non-composable is to assign it to `mutableStateOf` which
is not common. And in this case a user will be presented with two
options to apply quickfix and they would have to choose the right one.

Bug: 309364913
Test: AddComposableAnnotationQuickFixTest
Change-Id: I14dbf95721b0b6788c6e7b76172c98b713367753