Add isStartupTask to session start command and data models
The isStartupTask is a boolean flag passed in the BeginSession
command and is used/extracted from the resulting SessionStarted
data retrieved on successful session start. This allows for the
detection of whether the session/task is done on startup of a
process or not. If it is a startup task, then we will only invoke
a subset of the profiler task handler behavior.

Whether the task is a startup task or not will be stored in
StudioProfilers alongside the current task type. We can retrieve
this value from the SessionStarted data on successful session
start as aforementioned. The addition of whether the task is to
be started on startup or not has also been added via a boolean
field in the TaskHomeTabModel.

Bug: 277798531
Test: updated
Change-Id: Id8e8cfea7b67ebdfa505ea00a0575d640cc2d839