Drop assert check out of interest from ComposeModifierCompletionContributorTest
In the test, `extensionFunctionReturnsNonModifier` function returns
`Int`. For the completion of
`myWidgetWithModifier(modifier = Modifier.<caret>`, K1 recommends
`extensionFunctionReturnsNonModifier`, which is technically wrong
because `modifier` parameter needs an argument with `Modifier` type.

K1 relies on `KotlinCompletionContributor` that adds
`extensionFunctionReturnsNonModifier` as a lookup element. On the
other hand, K2 counterpart `KotlinFirCompletionContributor` does not
add it as a lookup element. Since the purpose of
ComposeModifierCompletionContributorTest is not testing
`KotlinCompletionContributor` and `KotlinFirCompletionContributor`
(regardless of the correctness), we drop the assert.

Bug: 284311057
Test: ComposeModifierCompletionContributorTest.modifierIsNotImported
Change-Id: Ieca371af93a83d08eb9cf2e16308e9b57add2202