Create resource quick fix: don't allow renaming
This quick fix is meant to let the user create a resource when they've
referenced it but it doesn't yet exist. If we let the user choose a
different name than what they've already typed, they end up with a
newly-created resource that doesn't actually fix the code they wrote.

It would be possible to allow the user to choose a new name, create the
newly-named resource, and then update the reference for them. But given
that this is used in a variety of code contexts (resource XML, kt, java)
and that it's comparatively easy to rename the resource afterwards, it
seems more prudent to just disallow renaming from the creation dialog.

Fixes: 318743410
Test: AndroidLayoutDomTest.testCreateResourceFromUsage
Change-Id: I9613ef93253808f680b908c82c61975316dffb3d