Add project context to ExplainSyncOrBuildOutput
This CL introduces a variety of prompting approaches in an attempt
to improve response quality for the "Ask Studio Bot" action available
from the build/sync tool window.

For errors with a description/details property, we include it in
the query but only use the shorter error name/message to represent
it in the chat UI.

For errors with a file location, we add the name and contents of that
file. If a line # is specified, it is pointed out with an arrow "-->".

For sync/Gradle-related errors, we include the contents of every Gradle
file in the project.

Evals: https://docs.google.com/spreadsheets/d/1r91WWsBL7U-1x8cy8lVckyhbuNfknnUWo9pebsJzL_E/edit?resourcekey=0-YUTFrUNzFUliZ-sX28DLKg#gid=0

While Studio Bot does not perform as well as Bard, and many of its
responses are bad or unhelpful, the context universally improves
the responses. Without context, the model can only hallucinate.

Bug: 310696988
Bug: 314179037
Test: included
Change-Id: Ibf5d8db59136832c6f54cab4dafe148ec0c8ba86