Fix MergedManifestInfo leak
Rather than listen for sync events for each MergedManifestInfo
instance, use the ProjectSyncModificationTracker and compare its value
when we need to know whether an info is out of date.

At present the ProjectSyncModificationTracker is incremented for every
syncEnded(), not just successful syncs.  This means that
MergedManifestInfo instances will be treated as out of date even when
a sync fails (and hence nothing about the project has in fact
changed).

Bug: 318986916
Test: manual, by inspection
Change-Id: I0d69fddc691bb7c69bfe6c3a8278fe56c42a2ef9