Make TaskArgs always non-null and pass into startTask
Previously, the createArgs method could return a null value for the
constructed TaskArgs. Now, because we will use TaskArgs for both
starting and loading a task (the two and only methods of performing
a task), createArgs will construct a non-null TaskArgs instance
always.

A refactor of the createArgs method is included in this change to
unify the logic of creating TaskArgs found in each task handler.

The TaskArgs constructed for starting the task will be used to
determine whether to start the task on startup or not. A later
change will make use of such TaskArgs in startTask.

Bug: 277797528
Bug: 277798531
Test: updated
Change-Id: I414570ec17288599a48bc8c0c081d73391a98e8c