Migrate GradleSpecificInitializer to AppLifecycleListener
It's not really a Gradle-specific initializer, though; at least some
of its actions are more general.  The install path checking is
important for essentially anything that interacts with jars, and the
JDK table cleanup is performed for Game Tools, which do not interact
with Gradle.  However, moving its operation to a defined point in the
lifecycle makes more of the application's facilities available.

The use of invokeLater[OnWriteThread] is also dubious, because the
asynchronous nature provides limited guarantees.  The presumed intent
is that the writes should complete before any project is opened, and
this *should* work presuming that write actions are processed in
order, but I can't actually find a strong guarantee anywhere.
Nevertheless, it is unlikely that the ordering desired here is
violated, and even if it is it would lead to the kind of symptoms that
are gone after an "Invalidate Caches and Restart" cycle.

Bug: 317992519
Test: manual
Change-Id: I1d4f1189ec709b513af29667548ef8c4cfaaa70c