Separate K2 analysis API use from template in compose completion
This commit moves the part of ComposeCompletionContributor relying on
the analysis APIs out of the template running on write action to avoid
the UI freeze.

Note that we still use the analysis API inside
ComposePostFormatProcessor, but we cannot avoid this. Since the template
gives freedom to users to update the expression whatever they want, we
cannot forecast the template result. In other words, we cannot run the
analysis API in advance. Therefore, we have to allow template to use
analysis API when we cannot avoid it like reformat and reference
shortener.

Bug: 300163247
Test: ComposeCompletionContributorTest
Change-Id: I3c6a2dd0cbe5b91f99310631d458072d8ff08876