Split LintBaseline issue into two
This is the IDE part of a change to split the "LintBaseline" issue ID
into two separate issues: one which reports whether any lint issues were
baselined (this continues to use the same issue id), and a new one,
"LintBaselineFixed", which reports whether there were lint issues in the
baseline which are no longer reported.

The reason for splitting these up is that users have wanted to configure
one of these into an error but not the other.

Test: Existing
Bug: Not filed
Change-Id: Ic8d5ecf409e4f231a250a800e4d03152827b07dd