Eliminate facets from `LaunchCompatibilityChecker`
Replace a pair of a facet and a callback taking the facet with a
callback without parameters. Also, make the callback parameter not null
as it is used as a not null one and is always passed at the caller side.

This is to simplify integration with other project systems.

Bug: n/a
Test: n/a
Change-Id: I912eef024d1c2fb016dfc8082aae0c0a5fa73a2d