Make the Merged Manifest Editor respond better to theme changes
Define and use lazy colors for the file color and the background,
meaning that a repaint will pick up the new theme automatically in
most cases.

Unfortunately, the details panel display is expressed in HTML, which
does not have the concept of a lazy color: the color is resolved to a
hex specification when the report is generated, not when it is
displayed.  There might somehow be a more sophisticated way to
implement late binding, for example by customizing and setting the
HTMLEditorKit's stylesheet, but the infrastructure for that seems poor
at present (there is one, shared stylesheet among all HTMLEditorKits,
so this could go very wrong), and so we instead expediently simply
regenerate the merging report.

Bug: 319070784
Test: manual
Change-Id: I9e948a8149cef85da8733f8eed0cfcaeb7622695