Fix flag logic: BUILD_OUTPUT_DOWNLOADS_INFORMATION
JetBrains commit 29237bd01c was not quite right. The problem is that
isBuildOutputShowsDownloadInfo() returns 'true' when running in
Android Studio, even if the default value of the
BUILD_OUTPUT_DOWNLOADS_INFORMATION flag is 'false' (assuming
the flag is not overridden by the user).

This is a follow-up to Change I2c5609a5b.

Change-Id: Id5c2d2ec61fba3d4e2df468092d3db4a6c613ce6