Register the listener immediately on construction
This is always called immediately after the listener is created,
but with a bunch of fancy logic to ensure it happens only once. By moving
this into the constructor (which we know will be called at most once per
project), we can keep this guarantee in a much simpler way.

Bug: 319332657
Test: N/A
Change-Id: Iec8c06b789b97467c80e60e7804b8ba0d95cba67