Store bindingLayoutGroups as CachedValue
This removes the need to track the updated modification count directly,
and allows us to get the benefits of the platforms optimizations around
cached values (and the extra safety of idempotency checks and whatnot).

The surrounding lock is still necessary since the result is also locally
cached in _bindingLayoutGroups, and that needs to happen with
appropriate ordering. b/322209412 tracks removing any handling of dumb
mode in this method after I've validated that the error line above is
never hit in stable builds, at which point the local caching and this
lock can be entirely removed.

Bug: n/a
Test: Covered by existing
Change-Id: I342212135d2bf76758914d01a036187aaf2c7629