[PSD] Filter flavor names from resolved models
If we are in the middle of an edit, we may have altered the dimension
of a product flavor.  Under those circumstances, we should prefer the
information from the parsed model, if it is available, in that it
reflects the intended state of the project; this means that we should
discard resolved flavor names of the same name when constructing the
keys for the product flavor collection.

Bug: 264979365
Test: manual
Change-Id: Ib8e9b116a78416645cff8a62ae5815c1ea8e8f50