Add FolderResourceRepository
This ResourceRepository allows for loading resources from the
plain file (compared to the one we currently have that requires
intellij VirtualFile). It is current purpose is to be used in the
the tests primarily to replace usages of AndroidTestCase where
ResourceRepository is required but there is no need for the full
project setup. However, the FolderResourceRepository is fully
functional ResourceRepository and therefore it is placed in the
production code with it's own tests.

Bug: 306121377
Test: FolderResourceRepositoryTest
Change-Id: I5ecc071829c116a837c73c831af065d9df425895