Remove use of Runnable in HtmlLinkManager
This also helps with fixing the project leak in the ShowExceptionFix
class. The use of Runnable encourages the capture of variables and
causes many leaks.
The main element used in the calls is the project which can be passed by
the HtmlLinkManager when the action is click avoiding the capture.

Fixes: 312909618
Test: StudioHtmlLinkManagerTest
Change-Id: If3453bff682c68f08de1f4e0db48c5747ba59e80