Deduplicate artifact in new session
In Pre-O devices,when a session is started and cpu capture is
taken and then when a new session is started and cpu capture
is taken, the old session cpu capture is carried to the new
session captures.

The cause of this behavior is that the cpu capture is not ended.
Setting 'isEnded' to true in event resolves the issue.

Bug: 309688086
Test: Modified
Change-Id: I0c4c0772f89af99d7767a03cfc5851014d6f7255