Do not store outdated NlModels in Visual Lint issues
Visual Lint issues store NlModels directly, and through NlComponents.
That is fine when those models exist in the design surface and
correspond to the preview that created the errors. However, when
stopping UI Check, the NlModels are disposed or reused for new previews.
In this case, it does not make sense to keep them, or rely on them to
perform tasks such as navigating to the origin of the error.

Bug: 312461091
Test: tests updated
Change-Id: I4c517479b2ff7bbe2fd034efeb9e132e4c135f10