Delete dead code
The StudioFlag to hide GUI managed adb server does not work. And likely
never worked because the UI elements are not in myPanel from where
they are remove()d. They are in a sub, unnamed, JPanel which contains all the elements.

There is no point fixing this code since these GUI elements have shipped
in 2020.

Test: NA
Bug: NA
Change-Id: I593951236edba4f418fcc82a34d0d000b1a3351d