Check for dumb mode in BindingScopeEnlarger
As far as I can tell from code analysis, existing bugs, and debugging,
`BindingScopeEnlarger` is the only caller of
`LayoutBindingModuleCache.bindingLayoutGroups` that calls while in dumb
mode. This change moves* local storages of the groups for usage in dumb
mode up to `BindingScopeEnlarger`, so that `LayoutBindingModuleCache`
doesn't have to.

*Technically this change adds storage a layer up, while also retaining
it in `LayoutBindingModuleCache`. This is done so as to not break
existing behavior in case there are other dumb mode callers that I
missed. The change of the log line at that location to an error means
that we'll start getting bugs files for those reports if indeed there is
such a caller.

Bug: 320471140, 322209412
Test: BindingScopeEnlargerTest.scopeDoesNotCacheStaleValuesInDumbMode
Change-Id: I48fe4b6df25a524b27196ff2a93800a29a70a370