Add LightBindingClass equality
LightBindingClass is stored in `CachedValue`s in various places. The
platform occasionally creates a `CachedValue` twice in order to validate
idempotency of the contained results. Since `LightBindingClass` doesn't
implement `equals()`, it always fails that check.

The new test was flaky last time I submitted and I reverted the change;
I've updated it and have run 500x to confirm it's no longer flaky:
https://fusion2.corp.google.com/invocations/94291743-f373-4345-8edc-871490034e12

Bug: 320471140
Test: LightBindingClassTest.testEquality
Change-Id: I2cba4b6be794e9b6ea1cc30cc5380fc3ca2ac6c5