Allow module selection only with transitive res
When offering to create a resource, a "context" module is provided. When
this code was written and transitive resources were the only option, it
made sense that a user might want to create the resource in the context
module, or in any of its dependent modules. (They often get to the
dialog by using the "create resource" quick fix on a non-resolving
`R.color.something` piece of code, and that `R` class allows access to
resources in the current module and all dependent modules.)

Now that non-transitive resources exist, it doesn't make sense to create
the resource on a different module. Doing so in the above example would
require changing the calling code to point to a different `R` class,
which is uninuitive and unnecessary.

Fixes: 318743105
Test: Manually verified
Change-Id: I0a90f81ba174dac29c355f504c77d47172d03922