Make new device selector work with BuildAndRunKMPTest
The existing device selector has a race condition that can cause E2E
tests to be flaky. The DeploymentTargetDevices are built by a series
of flow transformations, causing the deviceHandleFlow to be restarted
when one of its predecessors changes. It initially sends out an empty
list, which should quickly be replaced if there is a device. However,
the test invokes Run immediately following build completion, and since
build completion restarts this flow (due to dumb mode change), it
occasionally coincides with the "no devices" state.

This change moves the tracking of devices to a StateFlow that is
independent of other inputs; the final DeploymentTargetDevices are
produced by a lightweight process that combines DeviceHandleState with
the other inputs. This avoids the momentary loss of state of the old
implementation.

Bug: 322498468
Test: BuildAndRunKMPTest, 500 runs
Change-Id: Ice98eb30567597f19299ffa3a7fbc81875eef8c5