Make baseline RunConfiguration testCompileMode be ANDROID_TESTS
The baseline profile generator is effectively an on-device test (it is
run with AndroidJUnit4.class).  The testCompileMode field (from
PreferGradleMake) is probably mostly stale at this point; updating
this instance of it allows the correspondence with isTestConfiguration
to be clear.

Also simplify the declared superclasses of
AndroidBaselineProfileRunConfiguration: AndroidRunConfigurationBase
already implements many of the same interfaces.

Bug: n/a
Test: manual
Change-Id: Ibb08c61a329b0276b519b1f40afa4e1acb0d4c1b