Use isAndroidProject to determine creation of LiveEdit actions
Change I80abbedbf670e827c20c6ba5dee6bc2e7b46c7ce dispatches on the
presence or absence of a configured Android Sdk known to the IDE,
which would potentially enable the LiveEdit actions on non-Android
projects which happen to be opened in an IDE that has previously been
configured to open android projects.  Instead, return null if the
project associated with the editor itself is not an Android Project.

Bug: none filed
Test: none
Change-Id: Ie1e6cd307e10cfd608011a7ba4e82f0f3dd84f94