Make GradleVersionCatalogDetector a bit more EDT-safe
The maybeSuggestToml method, much like the AGP Upgrade Assistant's
maybeRecommendUpgrade, should do most of its work off the EDT -- in
particular for this case, we should not potentially compute the
DetectorResult on the EDT as part of it involves loading (and parsing)
a file to Psi.  Instead, perform the computation and notification on a
pooled thread.

One other use of the GradleVersionCatalogDetector result is when
preparing a sync event for metrics.  In that case, we cannot guarantee
that we are not on the EDT, and we also require synchronous operation.
Create a new property whose getter is required not to start a
computation, and a new DetectorResult to encapsulate that; use it in
GradleSyncEventLogger.

Annotate the various getters which might trigger computation with
@Slow.

Fixes: 322964389
Test: time will tell
Change-Id: Ib5a558f481676a0710d634d64f7ba8c4df64acc8