Change ConvertToComposeAction to use a text dialog
Now we display the resulting Compose code in a non-modal dialog, so
it's more convenient for users to visualize the resulting code, as well
as copy/paste as they see fit.

This change also fixes the disposable of the AndroidCoroutineScope
responsible for querying Studio Bot, which was using the project, but is
now using the dialog disposable.

Bug: 322759144
Test: TBD after the hackathon
Change-Id: I3c906043547577a8b38d54f5602a878f6bf8874a