Delay AAR cache invalidation until all events are processed
The change for bug 322954494 fixed half of the problem, but there is
still the symmetric half... In short, cache invalidation listeners in
IDEA happen to have dependencies between them, so the order is
important. However, they happen to contain TWO methods, one called
before the change and one after, and, naturally, the order is the same,
which means you cannot treat the listeners as DFS visitors on the
dependency graph between caches (which would require the "after" methods
to be called in reverse order of the "before" methods!) So it's easier
to just not create any more dependencies and only ever use these
listeners to publish more events into the queue, then handle these
events separately.

Bug: 323145834
Test: BuildVariantsIntegrationTest and OpenProjectIntegrationTest
Change-Id: I96fe9cba1356a0e39bd116e521ed753def215c47