Remove 'Apply' button in Task Settings
In task-based ux, Apply button in task settings pop-up doesn't
play a vital role and so is removed. Ok button does the saving of
values, and there is no major time consuming change in pop-up
that would need apply button (which can save each time the apply
button is clicked).

Bug: 323245305
Test: Added
Change-Id: I511f59acfde918a060c7e1e31fad4726ecd3e33a