Atrace config used for device < P
System trace task was disabled for device < P since perfetto
config was being returned on purpose relative to an existing bug
(b/313700658). Dependency to that bug has now been released (since
it's verified that atrace is working as expected in linux), and
thus this fix to use atrace is implemented.

Bug: 313701964
Test: Added
Change-Id: I3a70ded7d87d4ec7da3ebfbe0443a28238347bb5