Fix additional line breaks added to Compose output code
When parsing Studio Bot response, we were adding a line break for every
chunk received, even if we were not adding it to the final answer. As a
result, we ended up with multiple additional line breaks when reading
non-code chunks, e.g. citations.

This issue can be fixed by filtering out code chunks before applying the
joinToString function.

Bug: 322759144
Test: TBD after hackathon
Change-Id: If722afadc525065a8713e33e674f7a37952f8ecd