Disentangle AgpVersions from EmbeddedDistributionPaths
The two were coupled just for a single utility (effectively, just
through AGP_APP_PLUGIN_MARKER.  It's a bit of a shame to have to
duplicate the plugin marker data, but this allows moving the rest of
AgpVersions out of android core.  EmbeddedDistributionPaths, and some
of its related utilities such as GoogleMavenRepo, do still embed some
knowledge of AGP as a special case, but less than the rest of the
AgpVersions functionality.

Bug: 147586218
Test: existing
Change-Id: I7cc2e296f7d6decdc8d70fca1dddc50c653aa1ae