Use PreviewRefreshManager in CommonPreviewRepresentation
The change Iae1ca5791dee31e7f introduces
WearTilePreviewRepresentationTest. However, this test highlighted some
flakiness. There are two causes to the flakiness.

The first cause is a ConcurrentModificationException caused by adding
build listeners while iterating over existing build listeners in
TestProjectSystemBuildManager. This additional build listener is
added by setupListenerWhenSmartAndSynced when entering smart mode.

The second cause is from models that are either rendered twice, leading
to more previews rendered than expected, and from renders not
completing, leading to the preview representation always being in a
state of "refreshing" due to "isRefreshing" always returning true.

This CL fixes the second cause by using a PreviewRefreshManager which
handles cancelling extra render requests and fixes the "isRefreshing"
that gets stuck to true.

The first cause is fixed in the change Iae1ca5791dee31e7f by
ensuring we are in smart mode and synced before starting the build.

BYPASS_INCLUSIVE_LANGUAGE_REASON=the use of "dumb" is to refer to
the name the API uses

Bug: 322309354
Bug: 298966370
Bug: 322698408
Bug: 239802877
Test: manual and stress testing with 500 run iterations
  WearTilePreviewRepresentationTest which is introduced in the next CL
Change-Id: Iab16f642a7617c64a9f775c5cf13df07e9c744c9