Replace flaky waitForAnyPendingRefresh with StateFlow check
waitForAnyPendingRefresh was a refresh completion check mechanism
based on Completables. But due to how the refresh manager works,
this completables could be completed exceptionally (cancelled),
and in those cases some other refresh could still be running.

This change replaces such check with one based on the StateFlow
provided by the refresh manager.

Bug: N/A
Test: --runs_per_test=300 for all affected test classes
Change-Id: I3e8a5670c5970ed6fbc8785ebcc2cb138db57fdc