Hide dropdown when there are only glance/tiles previews
In Compose projects, we always consider that we have a Compose Preview
representation, even if there are no previews to show. This behavior
was introduced back when Compose Preview discoverability was a problem.
At that time, we only supported Compose and Custom View previews, and it
made sense to always display the Compose "empty panel", which explains
to users how to create a Compose Preview.

Now that we can also have Glance and Tile previews in Compose Projects,
we ended up with an undesirable UX when a file only have Glance or Tile
previews, because the "Compose" preview representation is always shown,
even if it will only display the empty panel.

In order to fix this issue, we need to calculate if the current file has
one or more Compose Previews before deciding if we want to consider the
corresponding preview representation. If there are no Compose Previews,
we will hide the representation. This means that if there are only Tile
or Glance previews in a file, the preview representation dropdown won't
be displayed, as it only causes confusion to users.

Fixes: 325444553
Test: TODO in a follow up
Change-Id: I13671a9f407139354cb92b426aa217021a66f07e