Add a non-DefaultActionGroup
The generation of our searchableOptions xml files for change detection
needs to map action classes to modules, and without an explicit class
declaration in the plugins.xml file, the action group is a direct
instance of DefaultActionGroup, which is then ascribed to the
`lib.app.jar` module.  Defining this empty class for the group holding
"Build APK(s)" and "Build Bundle(s)" allows it to be mapped to the
`android.jar` module instead.

Bug: n/a
Test: n/a
Change-Id: I096fc42590ba32b199e138e7d3b1cca4b34920b4