Show package-visible layouts in xml
b/111843761 and ag/5284600 initially introduced a check for a view to be
`public` in order to be resolved in XML layouts. That makes sense for
library references (or really any references in a different module), but
package-visible views in the current module are still usage. This change
loosens the restriction slightly.

Fixes: 318166201
Test: Included
Change-Id: I92d9b22f415f3f44e8602e8ec7962a78e5d2582f