Add support for split window
The RD tool window supports having multiple visble tabs at the
same time. This chage adapts embedded Layout Inspector to support
that.

In particular, since there can now be multiple ContentManagers
for the RD tool window, RunningDevicesStateObserver is updated
to keep track of all the existing ContentManagers and to look
through all of them when getting the state from RD.

Test: LayoutInspectorManagerTest
Bug: 324741151
Change-Id: Iaa00a4c22d873070e97b2d0a0211eb3b52267a85