[Toml] Look for the PsiElement in a readAction
This addresses the fact that reading the Psi must be done while
holding the read lock.  It's not clear to me why this doesn't then
trip up the logically next assertion, that slow operations should not
be done on the EDT: it seems that this at least could be a slow
operation.  But maybe the library toml file is always in caches?  At
least it has been in my manual tests.

Bug: 326192551
Test: existing, manual
Change-Id: I62c8ddac21008e9b60a42ed10ddbd46e2627561e