Fix property generation
Change: 25d7346e introduced a property value update when the same
components have new property values.

Unfortunately the check only looked at the property names which is
causing a problem when the selection is switching between 2 similar
components such as 2 TextFields. The property names will be the same
and since we are caching properties for each view the property values
are not updated.

Fixes: 326351646
Test: Unit test added.
Change-Id: I4303d8a4c4e511a44fdc762459770c59ce875fda