Handle no facet in LayoutBindingModuleCache
It's unclear to me how exactly the facet doesn't exist. Regardless, this was clearly introduced by ag/25923765, so I'll go back to what was there before.

I'm not able to write a test since the facet is always present for the scenarios I can come up with. Relying on code analysis (and the fact that this is what previously existed) to validate correctness.

Fixes: 326478863
Test: n/a
Change-Id: I6cd83e8269e4d226bda771ee1708ed09f0d74897