Have Glance and Tile preview elements test equality based
... on their values.

In CommonPreviewRepresentation, we do a full refresh when the rendered
elements are different from the preview elements. At the moment this
will always be the case even if the actual values on the preview
elements have not changed as we do not test equality based on the
preview elements' values.

This change changes the equality testing for preview elements so that it
is based on the values on the preview elements.

This change also updates WearTilePreviewElementModelAdapterTest and
GlancePreviewElementModelAdapterTest as the affinity calculation
produces an affinity of 0 for the same valued elements. Before the
preview elements were considered different due to the references being
different.

Bug: 326412365
Test: GlancePreviewElement, WearTilePreviewElement,
 GlancePreviewElementModelAdapterTest,
 WearTilePreviewElementModelAdapterTest,
 updated ComposePreviewElementModelAdapterTest,
 all other tests should still pass too.
Change-Id: I127f34dfce1127bb60fbe681d41b9ae61bf235b7