Add integration test for configuration cache failure
This particular failure can not happen on Sync (currently),
configuration cache is active only on build, but this is still
a valuable test of error messages generation and stats reporting.

Don't ignore listeners in replaced BuildViewManager same as it is done
for SyncViewManager. This is needed for build failure integration tests,
e.g. this test for configuration cache failure.

Bug: 324894199
Test: added ConfigurationCacheFailureTest
Change-Id: I9fb1122933fbbe1c46efb5b61af1278a44f15230