Replace stop recording icon in Task-Based UX
The non Task-Based UX utilizes two different stop icons:
one for stopping a session, which is used in the session
panel, and one for stopping a recording. The latter icon
is utilized in the allocation stage and live view, and
can be easily confused for a start recording button.

Because in the Task-Based UX, we do not need to differentiate
between stopping a session and a recording, we can simply
use the stop session icon, which is closer to a traditional
stop icon, to represent stopping the recoridng/live tasks.

Fixes: 326123991
Test: manual
Change-Id: I84b0066f56b7a1a19b41616f7ef482811e3646b4