Make ReRunUiCheckModeAction more robust
Until now, the action relied on the instance id of the preview to
decide which preview to restart UI Check mode on. That can fail when two
preview annotations are defined on the same preview, as they will both
have the same instance id. In addition, this couldn't detect that the
preview would still be available before attempting to restart the UI
Check.
With this change, the action uses the pointer to the preview definition
instead of the id. That means it can differentiate between two previews
defined on the same composable. In addition, it can react to changes in
the code, such as deleting the preview annotation.

Bug: N/A
Test: test updated
Change-Id: I30979da292a326efeab16379e380bf5fcde9e0c5