Fix deadlock during creation of AndroidRunConfigurations during tests
With IntelliJ 2024.1 a fix was added on DumbService.waitForSmartMode
commit 89c8425512420 addressing IDEA-347280 ("waitForSmartMode
doesn't wait") issue. This is causing deadlock when creating the
AndroidRunConfigurations after a successful sync, since the
waitForSmartMode will pause current thread, waiting until dumb mode
ends to be able to continue its execution something that will never
happen during tests because those are running synchronously.

This behaviour cannot be reproducible on prod since the code runs on
the background task. For this reason to align more with prod environment
and properly address the deadlock a new AndroidRunConfigurationsManager
class was added, responsible for creating the run configurations by using
Task.Backgroundable and store the operations under CompletableFuture,
allowing to wait for pending operations to finish during test execution.

Bug: 328860457
Test: ProjectStructureUsageTrackerTest
Change-Id: Ia3d9377c5b4cddb6eabd606232f683f2fb92c9d8