Pass layoutlib aware classloader
... to ParametrizedComposePreviewElementTemplate.

Parent classloader is normally the one that is able to load
Android framework classes (e.g. android.os.Parcelable). If we
provide a parent classloader that does not have layoutlib on its
classpath there is no way to load parameter classes that reference
Android framework classes. Therefore, we allow for specifying the
parent classloader so that the caller of
ParametrizedComposePreviewElementTemplate can specify the
classloader that is aware of Android framework classes.

Bug: 308641013
Test: manually verified
Change-Id: I33c571555287753d058cf46fee0aa23de5788ccf