Add parameterized previews to Compose Preview E2E test
Previews using a PreviewParameterProvider that references framework
classes were failing to render in release builds because these classes
could not be loaded.

We're adding a PreviewParameterProvider preview to the test project,
and will modify the ComposePreviewKotlin test in a follow up change
to render and verify this preview renders correctly to prevent future
regressions.

Bug: 328197204
Test: ComposePreviewKotlin
Change-Id: I3975950685dab0a0a74b232c48f9a37f7c7ca555