Improve disablement of startup tasks
...by disabling startup configurations in one place,
that being right after the consumption of the startup
configuration. This prevents re-use of startup configs
meant for prior tasks if the user invokes a profiler
toolbar action (e.g. Profile with low overhead/complete
data) after performing a task.

Also, before, the resetting of the selection state on
enter (as called by the onStartupTaskStart) invocation
was quite flaky (was supposed to be invoked on successful
start of startup task, but was not effectively doing so).
This change moves the call to reset the selection state
to when theres real confirmation that the startup task
has actually begun.

There was also a listener to FULL session changes that
was being invoked on startup task start, setting the
stage to a CpuProfilerStage or MainMemoryProfilerStage.
This logic needed to be flagged out as the startup task's
handler will already handle the setting of the stage.

Bug: 277798531
Fixes: 328310492
Test: manual
Change-Id: I0897f3d2cf080d57bde21cd9b33d130eeead1c3d