Lint: report quick-fix creation exceptions
Lint checks may throw exceptions when the IDE tries to get the
quick-fixes. This happens in LintExternalAnnotator.apply, where we are
trying to add annotations (source code highlights) for each problem
found by Lint. Previously, this would abort the processing of the
problems, and could lead to another exception being thrown because the
annotation (source code highlight) being created did not finish being
built. This meant we lost the context of the original exception.

We now catch such exceptions, report them via LOG.error, and complete
the building of the annotation (the quick-fixes will be missing). We
include the issue id in error messages.

Bug: 300908560
Test: Manual
Change-Id: I610f7df1651d8c5cb84edcdaf957848f6356a9ef