Allow exceptions to propagate from AVD launch
AvdManagerConnection returns a ListenableFuture<IDevice> when launching
an AVD. We do not want the IDevice, but we do want to know if launching
is aborted for other reasons. In particular, if the emulator process
terminates before connecting to adb, we need to stop showing the device
as transitioning. Thus, call await() on the future in the bridge code to
cause exceptions to be propagated to the device provisioner plugin.

Convert the exceptions thrown during AVD startup to
DeviceActionException so that they are treated as user-facing. Convert
the exceptions thrown by cancel operations to
DeviceActionCanceledException so that they do not get shown to the user
or logged as errors. Any "unexpected" exceptions (e.g.
NullPointerException) will not result in a popup and will be logged as
crashes.

Bug: 315691699
Test: in companion
Change-Id: I7224c28174e80ed9687e32314ad20265efa0927e