Use a ApplicationCoroutineScopeDisposable parent disposable
... when creating scope disposables through the
coroutineScope#scopeDisposable method.

The Job.invokeOnCompletion method does not have thread guarantees, cf
cf: https://github.com/Kotlin/kotlinx.coroutines/issues/3505. The error
was caused due to a race condition where the leak check in
UndisposedAndroidObjectsCheckerRule#checkUndisposedAndroidRelatedObjects
was run before the disposable was disposed inside
Job.invokeOnCompletion.

This problem was only really a problem in tests, as the code was always
executed. However, it is a problem for tests.

By using an application level service as a parent disposable, we ensure
that all the scopeDisposable Disposables are disposed when tearing down
the tests and before the leak checks.

Fixes: 328290264
Test: PreviewLifecycleManagerTest, ComposePreviewRepresentationTest,
  WearTilePreviewRepresentationTest
Change-Id: I4dbc72206d87bfef0886fac7929c2783adcb398b