Update the wear pairing device only for running FTL devices
Device templates do not have properties so we can not detect if the
device is paired from its information. With this change, we ensure that
we provide the pairing id only for running devices.

Additionally we change the device identifier for pairing to match the
wear pairing id.

Fixes: 326187840
Test: WearPairingManagerTest
Change-Id: I68f5bd77b7b7274bed4aa0b1a26fd94f3d3f9f6f
(cherry picked from commit 50ba6c80f544b4901f8cc5afccf27896ef884e90)