Use custom scope for DB light classes
Up until now, BindingScopeEnlarger has included light binding classes in
module search scopes by inspecting the underlying VirtualFiles. This is
error prone, since every time a new binding class is created (which
happens very often due to cache invalidation), there's a new VirtualFile
underlying the new class. Additionally, recent changes adding
`.equals()` to the binding class mean that two different light classes
(with two different VirtualFiles backing them) are considered
equivalent, but the same search scope will only recognize one of them.

I'm doing away with all of that by making the search scope instead look
for a marker indicating that a file is a light binding class (ie, an
object stored in the VirtualFile's user data). This is trivial to add
when the backing files are created, and easy to check from the search
scope. This marking object is stored in LayoutBindingModuleCache, which
means there's a different instance for each module; this allows search
scopes to be constructed with appropriate module- and dependent-level
visibility.

As a side benefit, this means that the search scope no longer needs to
be in 'smart' mode to figure out correct results, and we don't have the
possibility of stale results that used to be present.

Fixes: 320471140
Test: BindingScopeEnlargerTest
Change-Id: I59a075529c2babe7f8585a328542384c263f5b3a